requirements.txt
Flask==3.1.3
mysql-connector-python==9.1.0
gunicorn==22.0.0
Flask-Bcrypt==1.0.1
Flake8
bandit
python-dotenv
